new commit

+++ src/pages/hotel/Hotel.js
@@ -1,6 +1,5 @@
-'use client'
-
 import React, { useEffect, useState } from "react";
+import 'animate.css';
 import {
   HotelSection,
   SecContent,
@@ -20,6 +19,10 @@ import {
   ModalText,
   ModalTextSpan,
   ModalBtnClose,
+  ModalBtnEdit,
+  BtnPrev,
+  BtnNext,
+  BtnPrevNext,
 } from "../../styles/Hotel.Style";
 import { 
   Navbar2Container, 
@@ -44,20 +47,22 @@ import {
   faTrash,
 } from "@fortawesome/free-solid-svg-icons";
 import axios from "axios";
-import Image from "next/image";
+import Link from "next/link";
+import CreerHotel from "../../app/creerHotels/CreerHotel"
 
 const Hotel = () => {
   const [hotels, setHotels] = useState([]);
-  const [hotelsImg, setHotelsImg] = useState([]);
   const [nombre, setNombre] = useState(0);
   const [seeButtons, setSeeButtons] = useState(null);
   const [error, setError] = useState(null);
   const [showModal, setShowModal] = useState(false);
   const [selectedHotel, setSelectedHotel] = useState(null);
+  const [showCreateForm, setShowCreateForm] = useState(false);
+  const [text, setText] = useState('Créer un nouvel hôtel'); // Nouvel état pour le texte du bouton
+  const [currentHotelIndex, setCurrentHotelIndex] = useState(0); // État pour suivre l'index actuel de l'hôtel
 
   useEffect(() => {
     fetchData();
-    fetchData2();
   }, []);
 
   const fetchData = async () => {
@@ -71,18 +76,9 @@ const Hotel = () => {
     }
   };
 
-  const fetchData2 = async () => {
-    try {
-      const response = await axios.get("https://projetstage1backend.onrender.com/uploads/updImage");
-      setHotelsImg(response.data);
-    } catch (error) {
-      console.error('Erreur lors de la récupération des données img:', error);
-      setError('Erreur lors de la récupération des données img.');
-    }
-  };
-
   const handleSeeButton = (index, hotel) => {
     setSelectedHotel(hotel);
+    setCurrentHotelIndex(index); // Définir l'index de l'hôtel actuel
     setSeeButtons(index === seeButtons ? null : index);
   };
 
@@ -93,6 +89,27 @@ const Hotel = () => {
     } catch (error) {
       setError('Erreur lors de la suppression de l\'hôtel.');
       console.error('Delete error:', error);
+    }
+  };
+
+  const handleCreateButtonClick = () => {
+    setShowCreateForm(!showCreateForm);
+    setText(showCreateForm ? 'Créer un nouvel hôtel' : 'Fermer la modal');
+  };
+
+  const prevHotel = () => {
+    if (currentHotelIndex > 0) {
+      const prevIndex = currentHotelIndex - 1;
+      setSelectedHotel(hotels[prevIndex]);
+      setCurrentHotelIndex(prevIndex);
+    }
+  };
+
+  const nextHotel = () => {
+    if (currentHotelIndex < hotels.length - 1) {
+      const nextIndex = currentHotelIndex + 1;
+      setSelectedHotel(hotels[nextIndex]);
+      setCurrentHotelIndex(nextIndex);
     }
   };
 
@@ -108,16 +125,14 @@ const Hotel = () => {
               </Header2Title>
               <Header3Title>
                 <Header1Subtitle>
-                  <a href="/creerHotel">
-                    <ButtonModal>
-                      <HeaderButtonPlus>
-                        <StyleIconCreer>
-                          <FontAwesomeIcon icon={faPlus} size="1x" color="black" />
-                        </StyleIconCreer>
-                        <StyleSpanCreer>Créer un nouvel hôtel</StyleSpanCreer>
-                      </HeaderButtonPlus>
-                    </ButtonModal>
-                  </a>
+                  <ButtonModal onClick={handleCreateButtonClick}>
+                    <HeaderButtonPlus>
+                      <StyleIconCreer>
+                        <FontAwesomeIcon icon={faPlus} size="1x" color="black" />
+                      </StyleIconCreer>
+                      <StyleSpanCreer>{text}</StyleSpanCreer>
+                    </HeaderButtonPlus>
+                  </ButtonModal>
                 </Header1Subtitle>
               </Header3Title>
             </Flex2ColumnContainer>
@@ -126,54 +141,58 @@ const Hotel = () => {
       </Navbar2Container>
 
       {showModal && selectedHotel && (
-        <ModalDetails>
+        <ModalDetails className="animate__animated animate__bounce animate__backInDown">
           <ModalMere>
             <ModalTitle>{selectedHotel.nameHotel}</ModalTitle>
-              <div className="princ d-flex align-items-center justify-content-around">
-                  <div className="left">
-                    <ModalText>Adresse: 
-                        <ModalTextSpan>{selectedHotel.address}</ModalTextSpan>
-                      </ModalText>
-                      <ModalText>Email: 
-                        <ModalTextSpan>{selectedHotel.email}</ModalTextSpan>
-                      </ModalText>
-                      <ModalText>Numéro de téléphone: 
-                        <ModalTextSpan>{selectedHotel.number}</ModalTextSpan>
-                      </ModalText>
-                      <ModalText>Prix par nuit: 
-                        <ModalTextSpan>{selectedHotel.price}</ModalTextSpan>
-                    </ModalText>
-                  </div>
-                  <div className="right">
-                    {hotelsImg.map((img) => (
-                      <div key={img._id}>
-                        <img src={img.filename} alt={img.filename} width={300} height={200}/>
-                        <p>{img.nameImage}</p>
-                      </div>
-                    ))}
-                  </div>
+            <div className="princ d-flex align-items-center justify-content-around">
+              <div className="left">
+                <ModalText>Adresse: 
+                  <ModalTextSpan>{selectedHotel.address}</ModalTextSpan>
+                </ModalText>
+                <ModalText>Email: 
+                  <ModalTextSpan>{selectedHotel.email}</ModalTextSpan>
+                </ModalText>
+                <ModalText>Numéro de téléphone: 
+                  <ModalTextSpan>{selectedHotel.number}</ModalTextSpan>
+                </ModalText>
+                <ModalText>Prix par nuit: 
+                  <ModalTextSpan>{selectedHotel.price} {selectedHotel.devise}</ModalTextSpan>
+                </ModalText>
               </div>
+              <div className="right">
+                <img src={selectedHotel.image} alt={selectedHotel.filename} width={300} height={200}/>
+              </div>
+            </div>
           </ModalMere>
           <ModalBtnClose onClick={() => setShowModal(false)}>Fermer</ModalBtnClose>
+          <ModalBtnEdit>
+            <Link href={`/Edit?id=${selectedHotel._id}`}>Modifier</Link>
+          </ModalBtnEdit>
+          <BtnPrevNext>
+          <BtnPrev onClick={prevHotel}>Précédent</BtnPrev>
+          <BtnNext onClick={nextHotel}>Suivant</BtnNext>
+          </BtnPrevNext>
         </ModalDetails>
       )}
-  
-      {!showModal && (
+
+      {!showCreateForm && !showModal && (
         <SecContent>
           {hotels.map((hotel, index) => (
             <SingleDestination key={hotel._id} data-aos="fade-up">
               <ImageModalMere>
                 <TheBtns>
                   <TIcons onClick={() => handleSeeButton(index, hotel)}>
-                    <FontAwesomeIcon icon={faPlusCircle} color="black" />
+                    <FontAwesomeIcon icon={faPlusCircle} color="white" />
                   </TIcons>
                   {seeButtons === index && (
-                    <SeeAllButtons>
+                    <SeeAllButtons className="animate__animated animate__bounce animate__backInDown">
                       <TIcons onClick={() => handleDelete(hotel._id)}> 
                         <FontAwesomeIcon icon={faTrash} color="red" />
                       </TIcons>
-                      <TIcons>
-                        <FontAwesomeIcon icon={faEdit} color="yellow" />
+                      <TIcons onClick={() => handleSeeButton(index, hotel)}>
+                        <Link href={`/Edit?id=${hotel._id}`}>
+                          <FontAwesomeIcon icon={faEdit} color="yellow" />
+                        </Link>
                       </TIcons>
                       <TIcons onClick={() => setShowModal(true)}>
                         <FontAwesomeIcon icon={faEye} color="skyblue" />
@@ -182,25 +201,22 @@ const Hotel = () => {
                   )}
                 </TheBtns>
 
-                  <img src={hotel.image} alt={hotel.filename} width={300} height={200}/>
-
-                {/* {hotelsImg.map((img) => (
-                  <div key={img._id}>
-                    <img src={img.filename} alt={img.filename} width={300} height={200}/>
-                  </div>
-                ))} */}
+                <img src={hotel.image} alt={hotel.filename} width={300} height={200} />
               </ImageModalMere>
               <CardInfo>
                 <Continent>
                   <Address>{hotel.address}</Address>
                 </Continent>
                 <DestTitle>{hotel.nameHotel}</DestTitle>
-                <Price>{hotel.price} par nuit</Price>
+                <Price>{hotel.price} {hotel.devise} par nuit</Price>
               </CardInfo>
             </SingleDestination>
           ))}
         </SecContent>
       )}
+
+      {showCreateForm && <CreerHotel />} {/* Afficher le formulaire de création d'hôtel */}
+
     </HotelSection>
   );
 };
